% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_rate.R
\name{augment.functional_rate}
\alias{augment.functional_rate}
\title{Augment functional_rate}
\usage{
\method{augment}{functional_rate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"functional_rate"}.}

\item{...}{Additional arguments.}
}
\value{
A tibble of the full rate evaluation grid with predictions, derivatives,
and confidence intervals.
}
\description{
Augment functional_rate
}
