% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_rate.R
\name{augment}
\alias{augment}
\title{Augment generic for functional objects}
\usage{
augment(x, ...)
}
\arguments{
\item{x}{An object to augment.}

\item{...}{Additional arguments.}
}
\value{
An augmented data frame or tibble.
}
\description{
Augment generic for functional objects
}
