% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_pca.R
\name{augment_functional_pca}
\alias{augment_functional_pca}
\title{Augment functional PCA}
\usage{
augment_functional_pca(x)
}
\arguments{
\item{x}{An object of class \code{"r4pde_functional_pca"}.}
}
\description{
Augment functional PCA
}
