% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_rate.R
\name{augment_functional_rate}
\alias{augment_functional_rate}
\title{Augment functional_rate (alias)}
\usage{
augment_functional_rate(x, ...)
}
\arguments{
\item{x}{An object of class \code{"functional_rate"}.}

\item{...}{Additional arguments.}
}
\value{
A tibble of the full rate evaluation grid.
}
\description{
Augment functional_rate (alias)
}
