% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{BlastWheat}
\alias{BlastWheat}
\title{BlastWheat dataset}
\format{
A data frame with the following columns:
\describe{
\item{heading}{Date of heading}
\item{inc_mean}{Mean incidence}
\item{index_mean}{FHB index mean}
\item{latitude}{Latitude coordinate}
\item{location}{Experimental site name}
\item{longitude}{Longitude coordinate}
\item{state}{Brazilian state}
\item{study}{Study ID or code}
\item{year}{Crop year}
\item{yld_mean}{Mean yield}
}
}
\source{
Del Ponte Lab internal data
}
\usage{
BlastWheat
}
\description{
Wheat blast dataset with severity and weather covariates.
}
\keyword{datasets}
