\name{BudBlightSoybean}
\alias{BudBlightSoybean}
\docType{data}
\title{
Bud blight of soybean
}
\description{
Dataset of progress curves found on page 98 of “Study of Plant Disease Epidemics” (Madden et al. 2007). These curves depict the incidence of soybean plants showing symptoms of bud blight, which is caused by the tobacco streak virus. Four different treatments, corresponding to different planting dates, were evaluated using a randomized complete block design with four replicates. Each curve has four time-based assessments.
}
\usage{data("BudBlightSoybean")}
\format{
  A data frame with 64 observations on the following 4 variables.
  \describe{
    \item{\code{treat}}{a character vector}
    \item{\code{time}}{a numeric vector}
    \item{\code{block}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{

}
\source{
Madden, L. V., Hughes, G., & Van Den Bosch, F. (2007). The study of plant disease epidemics.
}
\references{

}
\examples{
data(BudBlightSoybean)

}
\keyword{datasets}
