% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{DidymellaWatermelon}
\alias{DidymellaWatermelon}
\title{DidymellaWatermelon dataset}
\format{
A data frame with:
\describe{
\item{EW_row}{Row position (east–west)}
\item{NS_col}{Column position (north–south)}
\item{dap}{Days after planting}
\item{severity}{Disease severity}
}
}
\source{
Del Ponte Lab internal data
}
\usage{
DidymellaWatermelon
}
\description{
Assessment of Didymella symptoms in watermelon plots.
}
\keyword{datasets}
