% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublets.test.R
\name{doublets.test}
\alias{doublets.test}
\title{Doublets Test}
\usage{
doublets.test(x)
}
\arguments{
\item{x}{A numeric vector representing the input data}
}
\value{
A list with elements Db (number of doublets), EDb (expected number of doublets), pvalue, and result (either "aggregation or clustering" or "randomness").
}
\description{
Perform a doublets test on the input data to test for aggregation or clustering or randomness.
}
\examples{
doublets.test(c(1, 0, 1, 1, 0, 1, 0, 0, 1, 1))

}
