% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSI.R
\name{DSI}
\alias{DSI}
\title{DSI}
\usage{
DSI(unit, class, max)
}
\arguments{
\item{unit}{A numeric vector representing the unit of measure}

\item{class}{A factor vector representing the class labels}

\item{max}{A numeric scalar representing the maximum possible value of the unit of measure}
}
\value{
A numeric scalar representing the DSI
}
\description{
Calculates the Disease Severity Index
}
\examples{
unit <- c(1:12)
class <- c(2,3,1,1,3,4,5,0,2,5,2,1)
ratings <- data.frame(unit, class)
DSI(unit = ratings$unit, class = ratings$class, max = 6)
}
