% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{FHBWheat}
\alias{FHBWheat}
\title{FHBWheat dataset}
\format{
A data frame with:
\describe{
\item{field}{Field identifier}
\item{i}{Row position}
\item{n}{Column position}
\item{quadrat}{Quadrat ID}
\item{season}{Crop season}
}
}
\source{
Del Ponte Lab internal data
}
\usage{
FHBWheat
}
\description{
Fusarium head blight quadrat assessments in wheat.
}
\keyword{datasets}
