% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{FHBWheatBrazil}
\alias{FHBWheatBrazil}
\title{FHBWheatBrazil dataset}
\format{
A data frame with the following columns:
\describe{
\item{study}{Study ID or code}
\item{year}{Crop year}
\item{location}{Experimental site name}
\item{state}{Brazilian state}
\item{lat}{Latitude coordinate}
\item{lon}{Longitude coordinate}
\item{cultivar}{Wheat cultivar}
\item{planting_date}{Planting date}
\item{flowering}{Flowering date}
\item{index}{FHB index}
}
}
\source{
Del Ponte Lab internal data
}
\usage{
FHBWheatBrazil
}
\description{
Fusarium head blight (FHB) index in wheat in Brazil.
}
\keyword{datasets}
