% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_pca.R
\name{functional_pca_clusters}
\alias{functional_pca_clusters}
\title{Cluster curves based on FPCA scores}
\usage{
functional_pca_clusters(
  x,
  k = NULL,
  components = NULL,
  method = c("kmeans", "hclust"),
  choose_k = c("none", "silhouette", "elbow"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{"r4pde_functional_pca"}.}

\item{k}{Number of clusters.}

\item{components}{Integer vector of components to use.}

\item{method}{Clustering method: "kmeans" or "hclust".}

\item{choose_k}{Method for suggesting k: "none", "silhouette", or "elbow".}

\item{...}{Additional arguments passed to clustering functions.}
}
\description{
Cluster curves based on FPCA scores
}
