\name{FusariumBanana}
\alias{FusariumBanana}
\docType{data}
\title{
Fusarium Wilt on Bananas
}
\description{
  Dataset on the location of diseased banana plants
}
\usage{data("FusariumBanana")}
\format{
  A data frame with 163 observations on the following 4 variables.
  \describe{
    \item{\code{lon}}{a numeric vector}
    \item{\code{lat}}{a numeric vector}
    \item{\code{marker}}{a character vector}
    \item{\code{field}}{a numeric vector}
  }
}
\details{
To do
}
\source{
To do
}
\references{
To do
}
\examples{
data(FusariumBanana)
## maybe str(FusariumBanana) ; plot(FusariumBanana) ...
}
\keyword{datasets}
