% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_pca.R
\name{get_fpca_scores}
\alias{get_fpca_scores}
\title{Get FPCA scores}
\usage{
get_fpca_scores(x, components = NULL)
}
\arguments{
\item{x}{An object of class \code{"r4pde_functional_pca"}.}

\item{components}{Optional integer vector of components to return.}
}
\description{
Get FPCA scores
}
