% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_pca.R
\name{get_fpca_variance}
\alias{get_fpca_variance}
\title{Get FPCA variance explained}
\usage{
get_fpca_variance(x)
}
\arguments{
\item{x}{An object of class \code{"r4pde_functional_pca"}.}
}
\description{
Get FPCA variance explained
}
