% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_functional_suppression_profiles_list.R
\name{plot.functional_suppression_profiles_list}
\alias{plot.functional_suppression_profiles_list}
\title{Plot a list of functional suppression profiles}
\usage{
\method{plot}{functional_suppression_profiles_list}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"functional_suppression_profiles_list"}.}

\item{...}{Additional arguments passed to \code{plot.functional_suppression_profiles}.}
}
\value{
A list of ggplot objects, one for each environment.
}
\description{
Plot a list of functional suppression profiles
}
