% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DSAF.R
\name{plot_DSAF}
\alias{plot_DSAF}
\title{Plot Dynamics and Structure Analysif of Foci (plot_DSAF)}
\usage{
plot_DSAF(df)
}
\arguments{
\item{df}{A dataframe containing at least three columns: 'x', 'y', and 'cluster_id'.
'x' and 'y' are spatial coordinates and 'cluster_id' is the cluster identifier
to which each cell belongs.}
}
\value{
A ggplot object with the scatter plot of foci (clusters).
}
\description{
This function creates a tile plot of the foci (cluster) identified by the DSAF function.
It colors each cell in a foci and labels the centroid of each cluster with the foci ID.
The 'ggplot2' package is used for the plot, and will be automatically installed if not already present.
}
\examples{
df <- data.frame(x = sample(1:100, 500, replace = TRUE),
                 y = sample(1:100, 500, replace = TRUE),
                 i = sample(0:1, 500, replace = TRUE, prob = c(0.7, 0.3)))

# Perform the DSAF
result <- DSAF(df)
# Plot the foci
plot_DSAF(result[[3]])

}
