% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functional_instability.R
\name{plot_functional_instability}
\alias{plot_functional_instability}
\title{Plot method for functional instability results}
\usage{
plot_functional_instability(object, type = c("overall", "space", "time"), ...)
}
\arguments{
\item{object}{Output from \code{\link[=functional_instability]{functional_instability()}}.}

\item{type}{One of \code{"overall"}, \code{"space"}, or \code{"time"}.}

\item{...}{Further arguments passed to methods.}
}
\value{
A ggplot object.
}
\description{
Plot method for functional instability results
}
