% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_rate.R
\name{print.functional_rate}
\alias{print.functional_rate}
\title{Print a functional_rate object}
\usage{
\method{print}{functional_rate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"functional_rate"}.}

\item{...}{Additional arguments.}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print a functional_rate object
}
