% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_resistance.R
\name{print.functional_resistance}
\alias{print.functional_resistance}
\title{Print functional_resistance}
\usage{
\method{print}{functional_resistance}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"functional_resistance"}.}

\item{...}{Additional arguments.}
}
\description{
Print functional_resistance
}
