% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functional_suppression_profiles.R
\name{print.functional_suppression_profiles}
\alias{print.functional_suppression_profiles}
\title{Print functional suppression profiles}
\usage{
\method{print}{functional_suppression_profiles}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"functional_suppression_profiles"}.}

\item{...}{Additional arguments.}
}
\description{
Print functional suppression profiles
}
