% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.r4pde_curve_diagnostics}
\alias{print.r4pde_curve_diagnostics}
\title{Print method for curve diagnostics}
\usage{
\method{print}{r4pde_curve_diagnostics}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"r4pde_curve_diagnostics"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for curve diagnostics
}
