% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_k.R
\name{print.suggest_k}
\alias{print.suggest_k}
\title{Print method for suggest_k}
\usage{
\method{print}{suggest_k}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"suggest_k"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for suggest_k
}
