% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functional_suppression_profiles.R
\name{print.summary.functional_suppression_profiles}
\alias{print.summary.functional_suppression_profiles}
\title{Print summary of functional suppression profiles}
\usage{
\method{print}{summary.functional_suppression_profiles}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.functional_suppression_profiles"}.}

\item{...}{Additional arguments.}
}
\description{
Print summary of functional suppression profiles
}
