% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_functional_suppression_profiles_list.R
\name{print.summary.functional_suppression_profiles_list}
\alias{print.summary.functional_suppression_profiles_list}
\title{Print summary of a list of functional suppression profiles}
\usage{
\method{print}{summary.functional_suppression_profiles_list}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.functional_suppression_profiles_list"}.}

\item{...}{Additional arguments.}
}
\description{
Print summary of a list of functional suppression profiles
}
