% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_dsp.R
\name{rank_dsp}
\alias{rank_dsp}
\title{Rank Treatments based on Functional DSP Descriptors}
\usage{
rank_dsp(
  summary_data,
  metrics = c("protected_area", "max_suppression", "persistence", "centroid"),
  higher_is_better = TRUE,
  group = NULL,
  ties.method = "average"
)
}
\arguments{
\item{summary_data}{The tibble output from \code{\link{functional_summary}}.}

\item{metrics}{Character vector of metrics to rank.}

\item{higher_is_better}{Logical vector or named logical vector indicating if higher values are better for each metric. If a single logical, applies to all.}

\item{group}{Optional character vector of grouping variables.}

\item{ties.method}{Character string specifying how to handle ties. Default is \code{"average"}.}
}
\value{
A tibble with ranks for each specified metric, and an \code{average_rank}.
}
\description{
Generates rankings of treatments based on extracted DSP functional descriptors.
}
