% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_pca.R
\name{reconstruct_curves}
\alias{reconstruct_curves}
\title{Reconstruct curves using specified FPCA components}
\usage{
reconstruct_curves(x, components = NULL)
}
\arguments{
\item{x}{An object of class \code{"r4pde_functional_pca"}.}

\item{components}{Integer vector of components to use for reconstruction. If NULL, uses all retained components.}
}
\description{
Reconstruct curves using specified FPCA components
}
