% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_dsp.R
\name{simulate_dsp_data}
\alias{simulate_dsp_data}
\title{#' Simulate Disease Suppression Profile Data}
\usage{
simulate_dsp_data()
}
\value{
A \code{tibble} with columns \code{treatment}, \code{rep}, \code{time}, \code{y_control}, \code{dsp_true}, \code{severity_true}, and \code{severity}.
}
\description{
Simulates a dataset for demonstrating functional analysis with Disease Suppression Profiles.
Returns a data frame with varying treatments and replicates over multiple time points.
}
