\name{SpatialAggregated}
\alias{SpatialAggregated}
\docType{data}
\title{
Simulated aggregated spatial xy data
}
\description{
A simulation of 200-point spatial xy data using the Shiny app https://delponte.shinyapps.io/spatial/
}
\usage{data("SpatialAggregated")}
\format{
  A data frame with 208 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
Simulation of 20 clusters with a standard deviation of 10 as defined in the ap
}
\source{
 https://delponte.shinyapps.io/spatial/
}

\examples{
data(SpatialAggregated)
## maybe str(SpatialAggregated) ; plot(SpatialAggregated) ...
}
\keyword{datasets}
