\name{SpatialRandom}
\alias{SpatialRandom}
\docType{data}
\title{
Simulated random spatial xy data
}
\description{
A simulation of 200-point spatial xy data using the Shiny app https://delponte.shinyapps.io/spatial/
}
\usage{data("SpatialRandom")}
\format{
  A data frame with 200 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
A random simulation
}
\source{
https://delponte.shinyapps.io/spatial/
}

\examples{
data(SpatialRandom)
## maybe str(SpatialRandom) ; plot(SpatialRandom) ...
}
\keyword{datasets}
