% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_rate.R
\name{summary.functional_rate}
\alias{summary.functional_rate}
\title{Summarize a functional_rate object}
\usage{
\method{summary}{functional_rate}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"functional_rate"}.}

\item{...}{Additional arguments.}
}
\value{
A tibble of curve-level rate phenotypes.
}
\description{
Summarize a functional_rate object
}
