% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functional_suppression_profiles.R
\name{summary.functional_suppression_profiles}
\alias{summary.functional_suppression_profiles}
\title{Summarize functional suppression profiles}
\usage{
\method{summary}{functional_suppression_profiles}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"functional_suppression_profiles"}.}

\item{...}{Additional arguments.}
}
\value{
A list of class \code{"summary.functional_suppression_profiles"}.
}
\description{
Summarize functional suppression profiles
}
