% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_functional_suppression_profiles_list.R
\name{summary.functional_suppression_profiles_list}
\alias{summary.functional_suppression_profiles_list}
\title{Summarize a list of functional suppression profiles}
\usage{
\method{summary}{functional_suppression_profiles_list}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"functional_suppression_profiles_list"}.}

\item{...}{Additional arguments.}
}
\value{
A list of summaries, one for each environment.
}
\description{
Summarize a list of functional suppression profiles
}
