% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_r4pde.R
\name{theme_r4pde}
\alias{theme_r4pde}
\title{Custom ggplot2 theme based on cowplot::theme_half_open}
\usage{
theme_r4pde(font_size = 16)
}
\arguments{
\item{font_size}{The base font size. Default is 16.}
}
\value{
A ggplot2 theme object.
}
\description{
This function creates a new ggplot2 theme by modifying the cowplot::theme_half_open theme.
It sets a custom font size and changes the panel background color to gray96.
}
\examples{
gg <- ggplot(mtcars, aes(x = mpg, y = wt)) +
  geom_point() +
  theme_r4pde()
print(gg)
}
