% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{BudBlightSoybean}
\alias{BudBlightSoybean}
\title{BudBlightSoybean dataset}
\format{
A data frame with the following columns:
\describe{
\item{block}{Block number}
\item{time}{Time point of assessment}
\item{treat}{Treatment name}
\item{y}{Incidence or severity value}
}
}
\source{
Del Ponte Lab internal data
}
\usage{
BudBlightSoybean
}
\description{
Soybean bud blight incidence in experimental blocks.
}
\keyword{datasets}
