\name{DidymellaWatermelon}
\alias{DidymellaWatermelon}
\title{Spatial data on gummy stem blight of watermelon}
\description{
 Data from the paper 'Temporal and spatial dynamics of watermelon gummy stem blight epidemics'.
}
\usage{data("DidymellaWatermelon")}
\format{
  A data frame with 1,344 observations on the following 4 variables.
  \describe{
    \item{\code{dap}}{An integer vector. Represents 'days after planting'.}
    \item{\code{NS_col}}{An integer vector. Represents the North-South column position.}
    \item{\code{EW_row}}{An integer vector. Represents the East-West row position.}
    \item{\code{severity}}{An integer vector. Represents the severity (0 to 4 scale) of gummy stem blight.}
  }
}
\details{
  more details
}
\source{
  Café-Filho, A.C., Santos, G.R. & Laranjeira, F.F. Temporal and spatial dynamics of watermelon gummy stem blight epidemics. Eur J Plant Pathol 128, 473–482 (2010). \url{https://doi.org/10.1007/s10658-010-9674-1}
}
\references{
  references
}
\examples{
  data(DidymellaWatermelon)
  head(DidymellaWatermelon)
}
\keyword{datasets}

