% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{FusariumBanana}
\alias{FusariumBanana}
\title{FusariumBanana dataset}
\format{
A data frame with:
\describe{
\item{field}{Field ID}
\item{lat}{Latitude}
\item{lon}{Longitude}
\item{marker}{Infection marker presence}
}
}
\source{
Del Ponte Lab internal data
}
\usage{
FusariumBanana
}
\description{
Observations of Fusarium symptoms in banana fields.
}
\keyword{datasets}
