% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oruns_test.R
\name{oruns_test}
\alias{oruns_test}
\title{Runs Test}
\usage{
oruns_test(x)
}
\arguments{
\item{x}{A numeric vector representing the input data}
}
\value{
A list with elements U (number of runs), EU (expected number of runs), pvalue, and result (either "clustering" or "randomness").
}
\description{
Perform a runs test on the input data to test for clustering or randomness.
}
\examples{
oruns_test(c(1, 0, 1, 1, 0, 1, 0, 0, 1, 1))

}
