% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oruns_test_byrowcol.R
\name{oruns_test_byrowcol}
\alias{oruns_test_byrowcol}
\title{Runs Test for Each Row and Column of a Binary Matrix}
\usage{
oruns_test_byrowcol(mat)
}
\arguments{
\item{mat}{A binary matrix (containing 0s and 1s, and possibly NAs).}
}
\value{
A list with four elements:
\item{row_results}{Data frame with test results for each row.}
\item{col_results}{Data frame with test results for each column.}
\item{row_summary}{Percentage summary of interpretation for rows.}
\item{col_summary}{Percentage summary of interpretation for columns.}
}
\description{
Applies the ordinary runs test to each row and column of a binary matrix individually.
}
\seealso{
\code{\link{oruns_test}}

Other Spatial analysis: 
\code{\link{AFSD}()},
\code{\link{BPL}()},
\code{\link{count_subareas}()},
\code{\link{count_subareas_random}()},
\code{\link{fit_gradients}()},
\code{\link{join_count}()},
\code{\link{oruns_test}()},
\code{\link{oruns_test_boustrophedon}()},
\code{\link{plot_AFSD}()}
}
\concept{Spatial analysis}
