% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_distances.R
\name{print.functional_distances}
\alias{print.functional_distances}
\title{Print functional_distances}
\usage{
\method{print}{functional_distances}(x, ...)
}
\description{
Print functional_distances
}
