% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_resistance.R
\name{print.functional_resistance}
\alias{print.functional_resistance}
\title{Print functional_resistance}
\usage{
\method{print}{functional_resistance}(x, ...)
}
\description{
Print functional_resistance
}
