% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{RustSoybean}
\alias{RustSoybean}
\title{RustSoybean dataset}
\format{
A data frame with:
\describe{
\item{detection}{Detection score or date}
\item{epidemia}{Epidemic phase}
\item{latitude}{Latitude}
\item{local}{Location name}
\item{longitude}{Longitude}
\item{planting}{Planting date or stage}
\item{severity}{Disease severity}
}
}
\source{
Del Ponte Lab internal data
}
\usage{
RustSoybean
}
\description{
Soybean rust severity and field metadata.
}
\keyword{datasets}
