% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{SpatialAggregated}
\alias{SpatialAggregated}
\title{SpatialAggregated dataset}
\format{
A data frame with:
\describe{
\item{x}{x-coordinate}
\item{y}{y-coordinate}
}
}
\source{
Simulated example
}
\usage{
SpatialAggregated
}
\description{
Simulated aggregated spatial binary disease pattern.
}
\keyword{datasets}
