% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{WhiteMoldSoybean}
\alias{WhiteMoldSoybean}
\title{WhiteMoldSoybean dataset}
\format{
A data frame with:
\describe{
\item{country}{Country name}
\item{elevation}{Field elevation}
\item{elevation_class}{Elevation class}
\item{harvest_year}{Year of harvest}
\item{inc}{Incidence}
\item{inc_check}{Check plot incidence}
\item{inc_class}{Incidence class}
\item{location}{Location name}
\item{region}{Geographical region}
\item{scl}{Soybean canopy layer}
\item{season}{Crop season}
\item{state}{State name}
\item{study}{Study identifier}
\item{treat}{Treatment applied}
\item{yld}{Yield}
\item{yld_check}{Yield of untreated check}
\item{yld_class}{Yield class}
}
}
\source{
Del Ponte Lab internal data
}
\usage{
WhiteMoldSoybean
}
\description{
National dataset of white mold severity and yield.
}
\keyword{datasets}
