% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowpane.R
\name{windowpane}
\alias{windowpane}
\title{Window Pane for Epidemiological Analysis}
\usage{
windowpane(
  data,
  end_date_col,
  date_col,
  variable,
  summary_type,
  threshold = NULL,
  window_lengths,
  direction = "backward",
  group_by_cols = NULL,
  date_format = "\%Y-\%m-\%d"
)
}
\arguments{
\item{data}{A dataframe containing the input data.}

\item{end_date_col}{The name of the column representing the end date.}

\item{date_col}{The name of the column representing the date variable.}

\item{variable}{The name of the column for which summary statistics are calculated.}

\item{summary_type}{The type of summary to calculate. Options are "mean", "sum",
"above_threshold", or "below_threshold".}

\item{threshold}{Optional numeric value used when summary_type is "above_threshold" or "below_threshold".}

\item{window_lengths}{A vector of window lengths (in days) for the calculations.}

\item{direction}{The direction of the window. Options are "backward" (default) or "forward".}

\item{group_by_cols}{Optional vector of column names for grouping the data.}

\item{date_format}{The format of the date columns. Default is "\%Y-\%m-\%d".}
}
\value{
A dataframe with the calculated summary values for each window.
}
\description{
This function calculates summary statistics within specified windows around a given end date
in a dataset, facilitating epidemiological analysis. It allows backward or forward window
calculations based on a user-defined variable and window lengths.
}
